feat: 포켓몬 정보 조회 API 추가 및 사용자 히스토리 컴포넌트 리팩토링
