+++ src/components/upload/hooks/useUploadMutation.ts
@@ -4,6 +4,7 @@ import { savePokemonResult } from '@/apis/imageResults.client';
 import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
+import { getPokemonInfo } from '@/apis/getPokemonInfo';
 
 export const useUploadMutation = () => {
   const router = useRouter();
@@ -17,9 +18,15 @@ export const useUploadMutation = () => {
       etag: string;
     }) => {
       const result = await matchPokemon(uploadedImgUrl); // 분석
-      const { id } = await savePokemonResult(etag, result); // 저장
+      const { text, genus } = await getPokemonInfo(result.matched_pokemon_id);
+      const pokemonInfo = {
+        ...result,
+        matched_pokemon_description: text,
+        matched_pokemon_genus: genus,
+      };
+      const { id } = await savePokemonResult(etag, pokemonInfo); // 저장
 
-      void saveUserHistory(etag, result, uploadedImgUrl).catch((e) => {
+      void saveUserHistory(etag, pokemonInfo, uploadedImgUrl).catch((e) => {
         if (process.env.NODE_ENV === 'development') {
           console.error('saveUserHistory error:', e);
         }
